# openapi.json
{"openapi":"3.0.2","info":{"title":"BetterAPI","description":"Introducing BetterAPI: Access and Utilize You.com with Ease\n\nBetterAPI is a free-to-use API designed to enhance your interaction with the You.com platform. With BetterAPI, you can seamlessly integrate YouChat into your projects and applications, providing a user-friendly and convenient way to leverage the capabilities of the chatbot.","version":"1.0.0","x-logo":{"url":"https://fastapi.tiangolo.com/img/logo-margin/logo-teal.png"}},"paths":{"/youchat":{"get":{"summary":"Youchat","operationId":"YouChat_youchat_get","parameters":[{"required":true,"schema":{"title":"Inputs"},"name":"inputs","in":"query"},{"required":true,"schema":{"title":"Key"},"name":"key","in":"query"},{"required":false,"schema":{"title":"Parameters","type":"string","default":""},"name":"parameters","in":"query"},{"required":false,"schema":{"title":"Contextid","default":""},"name":"contextid","in":"query"},{"required":false,"schema":{"title":"Powerpointer","default":false},"name":"powerpointer","in":"query"},{"required":false,"schema":{"title":"Stream","type":"boolean","default":false},"name":"stream","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"x-codeSamples":[{"lang":"Python","source":"import requests # import requests for the api call\n\nAPI_KEY = \"YOUR KEY HERE\" # your api key\n\nurl = \"https://api.betterapi.net/youchat?inputs=hello&key=\" + API_KEY # set api url\njson = requests.get(url).json() # load json form api\nprint(json[\"generated_text\"]) # print message response\n\n","label":"Python"}]},"post":{"summary":"Youchat","operationId":"YouChat_youchat_post","parameters":[{"required":false,"schema":{"title":"Parameters","default":{}},"name":"parameters","in":"query"},{"required":false,"schema":{"title":"Contextid","default":""},"name":"contextid","in":"query"},{"required":false,"schema":{"title":"Powerpointer","default":false},"name":"powerpointer","in":"query"}],"requestBody":{"content":{"application/json":{"schema":{"title":"Inputs"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/youdotcom/chat":{"get":{"summary":"New Endpoint","operationId":"new_endpoint_youdotcom_chat_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/betterapi/betterchat":{"get":{"summary":"Betterchat","operationId":"BetterChat_betterapi_betterchat_get","parameters":[{"required":true,"schema":{"title":"Message"},"name":"message","in":"query"},{"required":true,"schema":{"title":"Key"},"name":"key","in":"query"},{"required":false,"schema":{"title":"Contextid","default":""},"name":"contextid","in":"query"},{"required":false,"schema":{"title":"Powerpointer","default":false},"name":"powerpointer","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/google/bard":{"get":{"summary":"Google Bard","operationId":"Google_Bard_google_bard_get","parameters":[{"required":true,"schema":{"title":"Message"},"name":"message","in":"query"},{"required":true,"schema":{"title":"Key"},"name":"key","in":"query"},{"required":false,"schema":{"title":"Contextid","default":""},"name":"contextid","in":"query"},{"required":false,"schema":{"title":"Powerpointer","default":false},"name":"powerpointer","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/imagine/gen":{"get":{"summary":"Stable Diffusion","operationId":"Stable_Diffusion_imagine_gen_get","parameters":[{"required":true,"schema":{"title":"Prompt"},"name":"prompt","in":"query"},{"required":true,"schema":{"title":"Key"},"name":"key","in":"query"},{"required":false,"schema":{"title":"Style","default":"78"},"name":"style","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/generated/{id}/{name}":{"get":{"summary":"Image Endpoint","operationId":"image_endpoint_generated__id___name__get","parameters":[{"required":true,"schema":{"title":"Id","type":"string"},"name":"id","in":"path"},{"required":true,"schema":{"title":"Name","type":"string"},"name":"name","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/youdotcom/apps":{"get":{"summary":"You Apps","operationId":"You_Apps_youdotcom_apps_get","parameters":[{"required":true,"schema":{"title":"Search"},"name":"search","in":"query"},{"required":true,"schema":{"title":"Key"},"name":"key","in":"query"},{"required":false,"schema":{"title":"Appname","default":""},"name":"appname","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/youdotcom/history":{"get":{"summary":"History Out Of Date","operationId":"history_out_of_date_youdotcom_history_get","parameters":[{"required":true,"schema":{"title":"Key"},"name":"key","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/gen":{"get":{"summary":"Key Gen","operationId":"Key_gen_gen_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/youdocom/osapi":{"get":{"summary":"Osapi For You Dot Com","operationId":"osapi_for_you_dot_com_youdocom_osapi_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/youdotcom/cts":{"get":{"summary":"Youchat Text To Speach Out Of Date","operationId":"YouChat_Text_To_Speach_out_of_date_youdotcom_cts_get","parameters":[{"required":true,"schema":{"title":"Message"},"name":"message","in":"query"},{"required":true,"schema":{"title":"Key"},"name":"key","in":"query"},{"required":false,"schema":{"title":"Contextid","default":""},"name":"contextid","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/status":{"get":{"summary":"Read Items","operationId":"read_items_status_get","responses":{"200":{"description":"Successful Response","content":{"text/html":{"schema":{"type":"string"}}}}}}},"/":{"get":{"summary":"About","operationId":"about__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/betterchat":{"get":{"summary":"Betterchat Webui","operationId":"BetterChat_webui_betterchat_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/GeneratedPresentations/":{"get":{"summary":"Send Generated Ppt","operationId":"send_generated_ppt_GeneratedPresentations__get","parameters":[{"required":true,"schema":{"title":"Path"},"name":"path","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/powerpointer/GeneratedPresentations/":{"get":{"summary":"Send Generated Ppt2","operationId":"send_generated_ppt2_powerpointer_GeneratedPresentations__get","parameters":[{"required":true,"schema":{"title":"Path"},"name":"path","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/powerget":{"get":{"summary":"Powerget","operationId":"powerget_powerget_get","parameters":[{"required":true,"schema":{"title":"Msg"},"name":"msg","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}